Add demo components to wrap stories that need extra HTML
