+++ package.json
@@ -54,14 +54,14 @@
     "webpack-dev-server": "^4.15.1"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "^7.4.1",
-    "@storybook/addon-interactions": "^7.4.1",
-    "@storybook/addon-links": "^7.4.1",
+    "@storybook/addon-essentials": "^7.4.5",
+    "@storybook/addon-interactions": "^7.4.5",
+    "@storybook/addon-links": "^7.4.5",
     "@storybook/addon-svelte-csf": "^3.0.0",
-    "@storybook/blocks": "^7.4.1",
-    "@storybook/svelte": "^7.4.1",
-    "@storybook/svelte-webpack5": "^7.4.1",
-    "@storybook/testing-library": "^0.2.0",
+    "@storybook/blocks": "^7.4.5",
+    "@storybook/svelte": "^7.4.5",
+    "@storybook/svelte-webpack5": "^7.4.5",
+    "@storybook/testing-library": "^0.2.1",
     "eslint": "^7.32.0",
     "jest": "^29.6.3",
     "msw": "^1.2.3",
@@ -72,7 +72,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "serve": "^14.2.0",
-    "storybook": "^7.4.1"
+    "storybook": "7.4.5"
   },
   "scripts": {
     "build": "webpack",
