+++ src/addons/browser/stories/Filter.demo.svelte
@@ -1,0 +1,10 @@
+<script>
+  import Filter from "../Filter.svelte";
+  import Star from "../../../common/icons/Star.svelte";
+
+  export let args = {};
+</script>
+
+<Filter {...args}>
+  <Star slot="icon" />
+</Filter>
