+++ src/addons/browser/stories/Filter.stories.svelte
@@ -2,7 +2,7 @@
   import { Meta, Story, Template } from "@storybook/addon-svelte-csf";
 
   import Filter from "../Filter.svelte";
-  import Star from "../../../common/icons/Star.svelte";
+  import FilterDemo from "./Filter.demo.svelte";
 
   let args = { selected: false, name: "Pinned" };
 </script>
@@ -16,9 +16,7 @@
 />
 
 <Template let:args>
-  <Filter {...args}>
-    <Star slot="icon" />
-  </Filter>
+  <FilterDemo {args} />
 </Template>
 
 <Story name="Default" {args} />
