+++ src/addons/browser/stories/SearchInput.demo.svelte
@@ -1,0 +1,7 @@
+<script>
+  import SearchInput, { query } from "../SearchInput.svelte";
+</script>
+
+<SearchInput />
+
+<p>Query: {$query}</p>
