+++ src/addons/dispatch/stories/Form.stories.svelte
@@ -1,10 +1,10 @@
 <script>
   import { Meta, Story, Template } from "@storybook/addon-svelte-csf";
 
-  import Form, { values } from "../Form.svelte";
-  import * as addons from "../../fixtures/addons.json";
+  import Form from "../Form.svelte";
+  import FormDemo from "./Form.demo.svelte";
 
-  let form;
+  import * as addons from "../../fixtures/addons.json";
 </script>
 
 <Meta
@@ -14,21 +14,7 @@
 />
 
 <Template let:args={addon}>
-  <h2>{addon.name}</h2>
-
-  <Form
-    bind:this={form}
-    properties={addon.parameters.properties}
-    required={addon.parameters.required}
-    eventOptions={addon.parameters.eventOptions}
-  />
-
-  <div class="values">
-    <h2>Data</h2>
-    <code>
-      {JSON.stringify($values, null, 2)}
-    </code>
-  </div>
+  <FormDemo {addon} />
 </Template>
 
 <Story name="PDF Exporter" args={addons[0]} />
