+++ src/addons/sidebar/stories/Sidebar.demo.svelte
@@ -1,0 +1,15 @@
+<script>
+  import Sidebar from "../Sidebar.svelte";
+</script>
+
+<style>
+  .sidebar {
+    padding: 1.5rem 0;
+    width: var(--sidebar-width);
+    background-color: var(--sidebar);
+  }
+</style>
+
+<div class="sidebar">
+  <Sidebar />
+</div>
