+++ src/addons/stories/Drawer.stories.svelte
@@ -2,17 +2,10 @@
   import { Meta, Story, Template } from "@storybook/addon-svelte-csf";
 
   import Drawer from "../Drawer.svelte";
+  import DrawerDemo from "./Drawer.demo.svelte";
 
   let args = { visible: true, anchor: "right" };
 </script>
-
-<style>
-  .content {
-    padding: 0 1em;
-    min-width: 24em;
-    margin: 0 auto;
-  }
-</style>
 
 <Meta
   title="Add-Ons / Drawer"
@@ -22,12 +15,7 @@
 />
 
 <Template let:args>
-  <Drawer {...args}>
-    <div slot="content" class="content">
-      <h2>Drawer Contents</h2>
-      <p>Lorem ipsum dolor sit amet</p>
-    </div>
-  </Drawer>
+  <DrawerDemo {args} />
 </Template>
 
 <Story name="Righthand" {args} />
